backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
pydantic==2.5.3
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
websockets==12.0
opencv-python-headless==4.9.0.80
numpy==1.26.3
scipy==1.12.0
httpx==0.26.0
pytest==8.0.0
